+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Fix issue with major version upgrade initialization
  - Fix issue with PWM output dashboard element updating ([#517](https://github.com/kizniche/mycodo/issues/517))
+ - Fix dependency check for DS-type sensor calibration ([#518](https://github.com/kizniche/mycodo/issues/518))
 
 
 ## 6.2.2 (2018-08-22)